Fjerner EURES-deling i veilederapp
