+++ src/app/_common/utils/urlUtils.js
@@ -1,7 +1,7 @@
+import { getMiljø, Miljø } from "@/app/_common/utils/miljøUtils";
+
 export const navBaseUrl =
     process.env.NEXT_PUBLIC_ENVIRONMENT === "dev" ? "https://www.ansatt.dev.nav.no" : "https://www.nav.no";
 
 export const arbeidsplassenBaseUrl =
-    process.env.NEXT_PUBLIC_ENVIRONMENT === "dev"
-        ? "https://arbeidsplassen.intern.dev.nav.no"
-        : "https://arbeidsplassen.nav.no";
+    getMiljø() === Miljø.DEV ? "https://arbeidsplassen.intern.dev.nav.no" : "https://arbeidsplassen.nav.no";
